Add new functions

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: cterrasi <cterrasi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/21 16:56:48 by cterrasi          #+#    #+#             */
-/*   Updated: 2022/03/21 17:56:29 by cterrasi         ###   ########.fr       */
+/*   Updated: 2022/03/22 20:06:55 by cterrasi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 size_t	ft_strlen(char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*s++)
